actiz-document-analyzer/requirements/base.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Azure Services
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-storage-blob==12.19.0
openai==1.3.0

# File Processing
PyPDF2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
requests==2.31.0